tile collision

+++ src/object/Chest.java
@@ -5,14 +5,15 @@ import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.image.Image;
 import panel.GamePanel;
 
-public class Chest extends SuperObject{
-    public int z =1;
-    public Chest(double x,double y) {
+public class Chest extends SuperObject {
+    public int z = 1;
+
+    public Chest(int x, int y) {
         super();
-        name="Chest";
+        name = "Chest";
         image = new Image(ClassLoader.getSystemResourceAsStream("objects/normal_chest.png"));
-        setWorldX(x*gp.getTileSize());
-        setWorldY(y*gp.getTileSize());
+        setWorldX(x * gp.getTileSize());
+        setWorldY(y * gp.getTileSize());
     }
 
     @Override
